Bring the helpers into the relevant file
